ublox: parse +UGCNTRD stats as unsigned 64bit values
  [1551646332.583651] (ttyACM2): --> 'AT+UGCNTRD<CR>'
  [1551646332.626567] (ttyACM2): <-- '<CR><LF>+UGCNTRD: 1,0,0,0,0<CR><LF><CR><LF>+UGCNTRD: 2,1397316870,113728263578,1397316870,113728263578<CR><LF><CR><LF>OK<CR><LF>'
  [1551646332.627120] Reloading stats failed: Couldn't load primary PDP context 2 statistics: Error parsing session RX bytes

+++ plugins/ublox/mm-broadband-bearer-ublox.c
@@ -748,8 +748,8 @@ ugcntrd_ready (MMBaseModem  *modem,
     MMBroadbandBearerUblox *self;
     const gchar            *response;
     GError                 *error = NULL;
-    guint                   tx_bytes = 0;
-    guint                   rx_bytes = 0;
+    guint64                 tx_bytes = 0;
+    guint64                 rx_bytes = 0;
     guint                   cid;
 
     self = MM_BROADBAND_BEARER_UBLOX (g_task_get_source_object (task));
@@ -775,7 +775,6 @@ ugcntrd_ready (MMBaseModem  *modem,
         result->bytes_tx = tx_bytes;
         g_task_return_pointer (task, result, g_free);
     }
-
     g_object_unref (task);
 }
